Refactor statistics and types
Make Type::CloneType abstract. Change unsigned to size_t in some
places in namespace model. Add keyword override for overrided methods.
Refactor ToKeyValueStringPairs to ToKeyValueMap.Change CsvStats::GetMin
signature, remove GetMinMax_. Rename CsvStats::STDAndCentralMoment_ to
CsvStats::CalculateCentralMoment. Refactor CsvStats::ShowSample and
it's description. Add GetValue to TypedColumnData. Add IsOrdered static
method to Type. Add CreateColumnData static methods to FDAlgorithm.

+++ src/model/column.h
@@ -12,7 +12,7 @@ class Column {
 
 private:
     std::string name_;
-    unsigned int index_;
+    size_t index_;
     RelationalSchema const* schema_;
 
 public:
@@ -21,7 +21,7 @@ public:
         index_(index),
         schema_(schema) {}
 
-    unsigned int GetIndex() const { return index_; }
+    size_t GetIndex() const { return index_; }
 
     std::string GetName() const { return name_; }
     RelationalSchema const* GetSchema() const { return schema_; }
